refactor: Background Image. Eager load for top image for new pages (#2777)
* fix: image loading. optimised overside image.

* fix: remove unnecessary usebrowsersize

* fix: random <br> tag

* Redeploy

* fix: add alt

* fix: removed random div

* chore: removed unused components

* revert
